+++ src/library/occupancy_grid_map/point_cloud_obstacle.cc
@@ -49,11 +49,9 @@ void PointCloudObstacleTransform::GenerateLocalObstacleByLocalView(
   double slot_x_buffer = 1.0;
   double slot_y_buffer = 0.05;
   double safe_slot_width =
-      std::max(slot_width, static_cast<double>(config.max_car_width) + 0.1f) +
-      slot_y_buffer;
+      std::max(slot_width, config.max_car_width + 0.1) + slot_y_buffer;
   double safe_slot_length =
-      std::max(slot_length, static_cast<double>(config.car_length)) +
-      slot_x_buffer;
+      std::max(slot_length, config.car_length) + slot_x_buffer;
 
   slot_box.min_ = cdl::Vector2r(-0.5, -safe_slot_width / 2);
   slot_box.max_ = cdl::Vector2r(safe_slot_length, safe_slot_width / 2);
@@ -105,7 +103,7 @@ void PointCloudObstacleTransform::GenerateLocalObstacleByLocalView(
       // delete by ego
       if (config.astar_config.enable_delete_occ_in_ego) {
         gjk_.PolygonPointCollisionDetect(&is_collision, &ego_global_polygon,
-                                        Position2D(local.x, local.y));
+                                         Position2D(local.x, local.y));
 
         if (is_collision) {
           // ILOG_INFO << "xy " << local.x << " " << local.y
@@ -154,7 +152,7 @@ void PointCloudObstacleTransform::GenerateLocalObstacleByLocalView(
       // delete by ego
       if (config.astar_config.enable_delete_occ_in_ego) {
         gjk_.PolygonPointCollisionDetect(&is_collision, &ego_global_polygon,
-                                        Position2D(local.x, local.y));
+                                         Position2D(local.x, local.y));
 
         if (is_collision) {
           // ILOG_INFO << "xy " << local.x << " " << local.y << " delete
@@ -306,7 +304,7 @@ void PointCloudObstacleTransform::GenerateLocalObstacle(
   planning::PointCloudObstacle obs;
   for (auto& pair : obs_manager->GetObstacles()) {
     if (pair.second.GetObsMovementType() ==
-            apa_planner::ApaObsMovementType::MOTION) {
+        apa_planner::ApaObsMovementType::MOTION) {
       continue;
     }
 
@@ -318,8 +316,8 @@ void PointCloudObstacleTransform::GenerateLocalObstacle(
         continue;
       }
 
-      position.x = static_cast<double>(pt.x());
-      position.y = static_cast<double>(pt.y());
+      position.x = pt.x();
+      position.y = pt.y();
       if (config.astar_config.enable_delete_occ_in_ego) {
         gjk_.PolygonPointCollisionDetect(&is_collision, &ego_global_polygon,
                                          position);
@@ -351,45 +349,29 @@ void PointCloudObstacleTransform::GetCompactCarPolygonByParam(
     Polygon2D* box, const double lat_buffer, const double lon_buffer) {
   const apa_planner::ApaParameters& config = apa_param.GetParam();
 
-  box->vertexes[0].x =
-      static_cast<double>(config.car_vertex_x_vec[2]) + lon_buffer;
-  box->vertexes[0].y =
-      static_cast<double>(config.car_vertex_y_vec[2]) + lat_buffer;
+  box->vertexes[0].x = config.car_vertex_x_vec[2] + lon_buffer;
+  box->vertexes[0].y = config.car_vertex_y_vec[2] + lat_buffer;
 
-  box->vertexes[1].x =
-      static_cast<double>(config.car_vertex_x_vec[0]) + lon_buffer;
-  box->vertexes[1].y =
-      static_cast<double>(config.car_vertex_y_vec[0]) + lat_buffer;
+  box->vertexes[1].x = config.car_vertex_x_vec[0] + lon_buffer;
+  box->vertexes[1].y = config.car_vertex_y_vec[0] + lat_buffer;
 
-  box->vertexes[2].x =
-      static_cast<double>(config.car_vertex_x_vec[15]) - lon_buffer;
-  box->vertexes[2].y =
-      static_cast<double>(config.car_vertex_y_vec[15]) + lat_buffer;
+  box->vertexes[2].x = config.car_vertex_x_vec[15] - lon_buffer;
+  box->vertexes[2].y = config.car_vertex_y_vec[15] + lat_buffer;
 
-  box->vertexes[3].x =
-      static_cast<double>(config.car_vertex_x_vec[13]) - lon_buffer;
-  box->vertexes[3].y =
-      static_cast<double>(config.car_vertex_y_vec[13]) + lat_buffer;
+  box->vertexes[3].x = config.car_vertex_x_vec[13] - lon_buffer;
+  box->vertexes[3].y = config.car_vertex_y_vec[13] + lat_buffer;
 
-  box->vertexes[4].x =
-      static_cast<double>(config.car_vertex_x_vec[12]) - lon_buffer;
-  box->vertexes[4].y =
-      static_cast<double>(config.car_vertex_y_vec[12]) - lat_buffer;
+  box->vertexes[4].x = config.car_vertex_x_vec[12] - lon_buffer;
+  box->vertexes[4].y = config.car_vertex_y_vec[12] - lat_buffer;
 
-  box->vertexes[5].x =
-      static_cast<double>(config.car_vertex_x_vec[10]) - lon_buffer;
-  box->vertexes[5].y =
-      static_cast<double>(config.car_vertex_y_vec[10]) - lat_buffer;
+  box->vertexes[5].x = config.car_vertex_x_vec[10] - lon_buffer;
+  box->vertexes[5].y = config.car_vertex_y_vec[10] - lat_buffer;
 
-  box->vertexes[6].x =
-      static_cast<double>(config.car_vertex_x_vec[5]) + lon_buffer;
-  box->vertexes[6].y =
-      static_cast<double>(config.car_vertex_y_vec[5]) - lat_buffer;
+  box->vertexes[6].x = config.car_vertex_x_vec[5] + lon_buffer;
+  box->vertexes[6].y = config.car_vertex_y_vec[5] - lat_buffer;
 
-  box->vertexes[7].x =
-      static_cast<double>(config.car_vertex_x_vec[3]) + lon_buffer;
-  box->vertexes[7].y =
-      static_cast<double>(config.car_vertex_y_vec[3]) - lat_buffer;
+  box->vertexes[7].x = config.car_vertex_x_vec[3] + lon_buffer;
+  box->vertexes[7].y = config.car_vertex_y_vec[3] - lat_buffer;
 
   box->vertex_num = 8;
